% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Angat}
\alias{Angat}
\title{Angat Reservoir inflow}
\format{A data.frame with 3 variables
\describe{
\item{year}{From 1988 to 2014}
\item{month}{A factor with 12 levels from Jan to Dec}
\item{Q}{Inflow to the reservoir in cubic meter per second}
}}
\source{
\url{http://202.90.134.59/riverflow/01_avg30yr.aspx}
}
\usage{
Angat
}
\description{
Inflow to Angat Reservoir in cubic meter per second.
}
\keyword{datasets}
