% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{one_ONI_rep}
\alias{one_ONI_rep}
\title{ONI replicate}
\usage{
one_ONI_rep(num.years)
}
\arguments{
\item{num.years}{Number of years to simulate.}
}
\value{
A vector of the ONI time series.
}
\description{
Create one replicate of ONI time series using equation (4) in Burgers (1999) and parameter values reported in the same paper.
}
