% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ONI}
\alias{ONI}
\title{Oceanic ENSO Index}
\format{A data.table with 3 variables
\describe{
\item{year}{From 1950 to 2015}
\item{month}{A factor with 12 levels from Jan to Dec. Each month is the average of 3 consecutive months, i.e, Jan denotes DJF average}
\item{ONI}{Oceanic ENSO Index}
}}
\source{
NOAA \url{http://origin.cpc.ncep.noaa.gov/products/analysis_monitoring/ensostuff/ONI_v5.php}
}
\usage{
ONI
}
\description{
Oceanic ENSO Index
}
\keyword{datasets}
