% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{ONI_reps}
\alias{ONI_reps}
\title{ONI replicates}
\usage{
ONI_reps(ONI.obs, years = NULL, num.reps = 100, trim = NULL,
  var.cor = FALSE)
}
\arguments{
\item{ONI.obs}{A data frame of observed ONI with three columns (year, month, ONI). By default, the replicates will have the same years as ONI.obs, unless \code{years} is provided.}

\item{years}{If provided, the replicates will have these years.}

\item{num.reps}{Number of replicates to be generated. Default is 100.}

\item{var.cor}{Whether variance correction is performed. Setting \code{var.cor = TRUE} will ignore \code{norm.factor}.}

\item{norm.factor}{Used to cut off values that are too high or too low. Values that are more than \code{norm.factor * max(abs(obs))} will be set to \code{max(abs(obs))}. Default is 2.}
}
\value{
A data.table of 4 columns (rep, year, month, ONI).
}
\description{
Generate multiple stochastic replicates of ONI time series according to Burgers (1999).
}
