% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_reps}
\alias{plot_reps}
\title{Plot stochastic replicates}
\usage{
plot_reps(reps, obs, var.name)
}
\arguments{
\item{reps}{Stochastic replicate}

\item{obs}{Observed time series}

\item{var.name}{Name of the variable column to be plotted, must be the same in reps and obs}
}
\value{
A ggplot2 plot
}
\description{
Diagnostic plot to check the stochastic replicates by comparing it with the observed time series
}
