% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{Q_ARMAX}
\alias{Q_ARMAX}
\title{Streamflow replicates with ARMAX}
\usage{
Q_ARMAX(Q.obs, model, X = NULL, X.name = NULL, years = NULL,
  num.reps = 100, trim = NULL, trans = "log")
}
\arguments{
\item{model}{A time series model, e.g., that generated by \code{forecast::Arima()}.}

\item{X}{Exogeneous input, a data.table with 4 columns (rep, year, month, X). If there are more than 4 columns, the desirable column name to be used as exogenous input should be given in \code{X.name}; otherwise the function will take the 4th column by default. If provided, \code{num.reps} will be ignored.}

\item{years}{If provided, the replicates will have these years.}

\item{num.reps}{Number of replicates to be generated. Default is 100.}

\item{ONI.obs}{A data frame of observed streamflow with three columns (year, month, Q). By default, the replicates will have the same years as Q.obs, unless \code{years} is provided.}
}
\description{
Generate stochastic replicates of streamflow with the ARMAX model. If exogneous inputs are not provided then the ARMA model will be used.
}
