% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{Q_reps}
\alias{Q_reps}
\title{Stochastic streamflow replicates}
\usage{
Q_reps(Q.obs, method, order, X.obs = NULL, X.reps = NULL,
  X.name = NULL, X.lag = 0, years = NULL, num.reps = 100,
  trim = 2, Q.trans = "log-std", X.trans = "std")
}
\arguments{
\item{Q.obs}{A data frame of observed streamflow with three columns (year, month, Q). By default, the replicates will have the same years as Q.obs, unless \code{years} is provided.}

\item{method}{Either 'arima' or 'arimax'. 'parma' will be included soon.}

\item{order}{A vector of ARIMA model order. See \code{forecast::Arima} for details.}

\item{X.obs}{Observed exogeneous input, a data frame with 3 columns (year, month, X). If there are more than 3 columns, the desirable column name to be used as exogenous input should be given in \code{X.name}; otherwise the function will take the 3th column by default.}

\item{X.reps}{Replicates of X. See \link{ONI_reps}.}

\item{X.name}{String, the name of the column used to represent the exogenous input.}

\item{X.lag}{Integer, time lag of X for model fitting}

\item{years}{If provided, the replicates will be indexed by these years.}

\item{num.reps}{Number of replicates to be generated. Default is 100.}

\item{trim}{Used to cut off values that are too high or too low. Values that are more than \code{trim * max(abs(obs))} will be set to \code{max(abs(obs))}. Default is 2. Supply \code{trim = NULL} if you don't want trimming.}

\item{Q.trans}{The pre-processing transformation that will be applied to Q.obs before fitting.
\describe{
\item{log}{Log-transformation.}
\item{std}{Standardization: substracting monthly mean and dividing by monthly standard deviation.}
\item{log-std}{First take log-transform, then do standardization.}
}}

\item{X.trans}{Same as \code{Q.trans} but applied on \code{X.obs} and \code{X.reps}.}
}
\value{
A data.table with the following collumns: year, month, X (if \code{method == 'armax'}) and Q.
}
\description{
Generate stochastic replicates of monthly streamflow time series.
}
\details{
First, an ARMA or ARMAX model will be fitted using the overlapping period of \code{Q.obs} and \code{X.obs}. Then, an ARMA model will be fitted using \code{forecast::Arima}. Finally, stochastic time series will be generated using \code{stats::simulate}.
}
