% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shy-package.R
\docType{package}
\name{shy}
\alias{shy}
\alias{shy-package}
\title{shy: Stochastic Hydrology}
\description{
Utility functions for stochastic hydrological analyses
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ntthung/shy}
  \item Report bugs at \url{https://github.com/ntthung/shy/issues}
}

}
\author{
\strong{Maintainer}: Hung Nguyen \email{ntthung@gmail.com}

}
\keyword{internal}
